changed package structure
-> changed package structure
-> removed the test controller class which are used for the test purpose
