+++ logs/scoutCompass.log
@@ -17,6 +17,6 @@
 2024-02-11T14:50:37.531+05:30  INFO 11860 --- [main] o.h.e.t.j.p.i.JtaPlatformInitiator       : HHH000489: No JTA platform available (set 'hibernate.transaction.jta.platform' to enable JTA platform integration)
 2024-02-11T14:50:38.003+05:30  INFO 11860 --- [main] j.LocalContainerEntityManagerFactoryBean : Initialized JPA EntityManagerFactory for persistence unit 'default'
 2024-02-11T14:50:38.549+05:30  WARN 11860 --- [main] JpaBaseConfiguration$JpaWebConfiguration : spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning
-2024-02-11T14:50:38.839+05:30  INFO 11860 --- [main] o.s.s.web.DefaultSecurityFilterChain     : Will secure any request with [org.springframework.security.web.session.DisableEncodeUrlFilter@3fe21d0e, org.springframework.security.web.context.request.async.WebAsyncManagerIntegrationFilter@787b7796, org.springframework.security.web.context.SecurityContextHolderFilter@73483d4b, org.springframework.security.web.header.HeaderWriterFilter@69e308e2, org.springframework.web.filter.CorsFilter@1c512bb, org.springframework.security.web.authentication.logout.LogoutFilter@121c1a08, com.scoutcomapss.api.security.config.JwtAuthenticationFilter@36e358de, org.springframework.security.web.savedrequest.RequestCacheAwareFilter@67eabe02, org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestFilter@4197e859, org.springframework.security.web.authentication.AnonymousAuthenticationFilter@a0e35c3, org.springframework.security.web.session.SessionManagementFilter@682270ff, org.springframework.security.web.access.ExceptionTranslationFilter@322593ed, org.springframework.security.web.access.intercept.AuthorizationFilter@77185a2]
+2024-02-11T14:50:38.839+05:30  INFO 11860 --- [main] o.s.s.web.DefaultSecurityFilterChain     : Will secure any request with [org.springframework.security.web.session.DisableEncodeUrlFilter@3fe21d0e, org.springframework.security.web.context.request.async.WebAsyncManagerIntegrationFilter@787b7796, org.springframework.security.web.context.SecurityContextHolderFilter@73483d4b, org.springframework.security.web.header.HeaderWriterFilter@69e308e2, org.springframework.web.filter.CorsFilter@1c512bb, org.springframework.security.web.authentication.logout.LogoutFilter@121c1a08, JwtAuthenticationFilter@36e358de, org.springframework.security.web.savedrequest.RequestCacheAwareFilter@67eabe02, org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestFilter@4197e859, org.springframework.security.web.authentication.AnonymousAuthenticationFilter@a0e35c3, org.springframework.security.web.session.SessionManagementFilter@682270ff, org.springframework.security.web.access.ExceptionTranslationFilter@322593ed, org.springframework.security.web.access.intercept.AuthorizationFilter@77185a2]
 2024-02-11T14:50:39.067+05:30  INFO 11860 --- [main] o.s.b.w.embedded.tomcat.TomcatWebServer  : Tomcat started on port 8081 (http) with context path ''
 2024-02-11T14:50:39.083+05:30  INFO 11860 --- [main] c.s.api.ScoutCompassApplication          : Started ScoutCompassApplication in 6.334 seconds (process running for 8.067)
