+++ src/main/java/com/scoutcomapss/api/auth/AuthenticationController.java
@@ -1,4 +1,4 @@
-package com.scoutcomapss.api.security.auth;
+package com.scoutcomapss.api.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
