+++ src/main/java/com/scoutcomapss/api/auth/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.scoutcomapss.api.security.auth;
+package com.scoutcomapss.api.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationResponse {
-  private String token;
+public class AuthenticationRequest {
+  private String email;
+  private String password;
 }
