+++ src/main/java/com/scoutcomapss/api/auth/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package com.scoutcomapss.api.security.auth;
+package com.scoutcomapss.api.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationRequest {
-  private String email;
-  private String password;
+public class AuthenticationResponse {
+  private String token;
 }
