+++ src/main/java/com/scoutcomapss/api/auth/ScoutRegisterRequest.java
@@ -1,6 +1,5 @@
-package com.scoutcomapss.api.security.auth;
+package com.scoutcomapss.api.auth;
 
-import com.scoutcomapss.api.user.Instructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
