+++ src/main/java/com/scoutcomapss/api/auth/user/Instructor.java
@@ -1,15 +1,12 @@
-package com.scoutcomapss.api.user;
+package com.scoutcomapss.api.auth.user;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Date;
 import java.util.List;
