+++ src/main/java/com/scoutcomapss/api/auth/user/Role.java
@@ -1,4 +1,4 @@
-package com.scoutcomapss.api.user;
+package com.scoutcomapss.api.auth.user;
 
 public enum Role {
