+++ src/main/java/com/scoutcomapss/api/auth/user/Scout.java
@@ -1,17 +1,12 @@
-package com.scoutcomapss.api.user;
+package com.scoutcomapss.api.auth.user;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.Collection;
 import java.util.Date;
-import java.util.List;
 
 @Data
 @Builder
